move intermediate_outputs to trace data
