+++ mlflow/entities/trace.py
@@ -12,7 +12,6 @@ from mlflow.entities.trace_data import TraceData
 from mlflow.entities.trace_info import TraceInfo
 from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE
-from mlflow.tracing.constant import SpanAttributeKey
 
 _logger = logging.getLogger(__name__)
 
@@ -232,23 +231,3 @@ class Trace(_MlflowObject):
             "spans",
             "tags",
         ]
-
-    @property
-    def intermediate_outputs(self) -> Optional[dict[str, Any]]:
-        """
-        Returns intermediate outputs within the trace.
-        There are mainly two flows to return intermediate outputs:
-        1. When a trace only has one root span,
-        return `intermediate_outputs` attribute of the span.
-        2. When a trace is created normally with a tree of spans,
-        aggregate the outputs of non-root spans.
-        """
-        root_span = self._get_root_span()
-        if root_span and root_span.get_attribute(SpanAttributeKey.INTERMEDIATE_OUTPUTS):
-            return root_span.get_attribute(SpanAttributeKey.INTERMEDIATE_OUTPUTS)
-        # TODO: handle the second case for a normal trace with spans
-
-    def _get_root_span(self) -> Optional[Span]:
-        for span in self.data.spans:
-            if span.parent_id is None:
-                return span
