+++ mlflow/entities/trace_data.py
@@ -2,6 +2,7 @@ from dataclasses import dataclass, field
 from typing import Any, Optional
 
 from mlflow.entities import Span
+from mlflow.tracing.constant import SpanAttributeKey
 
 
 @dataclass
@@ -36,3 +37,23 @@ class TraceData:
             "request": self.request,
             "response": self.response,
         }
+
+    @property
+    def intermediate_outputs(self) -> Optional[dict[str, Any]]:
+        """
+        Returns intermediate outputs within the trace.
+        There are mainly two flows to return intermediate outputs:
+        1. When a trace only has one root span,
+        return `intermediate_outputs` attribute of the span.
+        2. When a trace is created normally with a tree of spans,
+        aggregate the outputs of non-root spans.
+        """
+        root_span = self._get_root_span()
+        if root_span and root_span.get_attribute(SpanAttributeKey.INTERMEDIATE_OUTPUTS):
+            return root_span.get_attribute(SpanAttributeKey.INTERMEDIATE_OUTPUTS)
+        # TODO: handle the second case for a normal trace with spans
+
+    def _get_root_span(self) -> Optional[Span]:
+        for span in self.spans:
+            if span.parent_id is None:
+                return span
